+++ Sources/PopupBridge/POPPopupBridgeSwift.swift
@@ -4,20 +4,23 @@ import SafariServices
 @objcMembers
 public class POPPopupBridgeSwift: NSObject, WKScriptMessageHandler, SFSafariViewControllerDelegate {
     
-    // Constants
-    let kPOPScriptMessageHandlerName = "POPPopupBridge"
-    let kPOPURLHost = "popupbridgev1"
+    // MARK: - Private Properties
     
-    // Properties
-    let webView: WKWebView
-    let delegate: POPPopupBridgeDelegateSwift
-    static var scheme: String?
-    var safariViewController: SFSafariViewController?
+    private let messageHandlerName = "POPPopupBridge"
+    private let hostName = "popupbridgev1"
     
+    private let webView: WKWebView
+    private let delegate: POPPopupBridgeDelegateSwift
+    private var safariViewController: SFSafariViewController?
+    
+    private static var scheme: String?
     private static var returnBlock: ((URL) -> Bool)?
     
-    // TODO: - make unfailable & require scheme in init, versus injecting via static method.
-    @objc public init?(webView: WKWebView, delegate: POPPopupBridgeDelegateSwift) {
+    // MARK: - Initializer
+    
+    // TODO: - make unfailable
+    public init?(webView: WKWebView, delegate: POPPopupBridgeDelegateSwift) {
+        // TODO: - require scheme in init, versus injecting via static method.
         guard let scheme = POPPopupBridgeSwift.scheme else {
             let exception = NSException(
                 name: NSExceptionName(rawValue: "POPPopupBridgeSchemeNotSet"),
@@ -27,32 +30,16 @@ public class POPPopupBridgeSwift: NSObject, WKScriptMessageHandler, SFSafariView
             return nil
         }
         
-        // Step 1 - get merchant's webView
         self.webView = webView
-        self.delegate = delegate // merchant implements POPUpBridgeDelegate
+        self.delegate = delegate
         
-        super.init() // do we need this?
+        super.init()
         
-        // Step 2 - construct a POPMessageDelegate
-        webView.configuration.userContentController.add(self, name: kPOPScriptMessageHandlerName)
+        configureWebView(scheme: scheme)
         
-        // Step 3 - Create javascript code and inject into WebView
-        let javascript = UserScript(
-            scheme: scheme,
-            scriptMessageHandlerName: kPOPScriptMessageHandlerName,
-            host: kPOPURLHost
-        ).rawJavascript
-        
-        let script = WKUserScript(
-            source: javascript,
-            injectionTime: .atDocumentStart,
-            forMainFrameOnly: true
-        )
-        webView.configuration.userContentController.addUserScript(script)
-        
-        // Handle URL from pop up
+        // TODO: - Remove returnBlock definition from init
         POPPopupBridgeSwift.returnBlock = { (url : URL) -> Bool in
-            guard let script = self.parseResponseFromPopUpToPassBacktoWebView(url: url) else {
+            guard let script = self.constructJavaScriptCompletionResult(returnURL: url) else {
                 return false
             }
             
@@ -61,11 +48,17 @@ public class POPPopupBridgeSwift: NSObject, WKScriptMessageHandler, SFSafariView
         }
     }
     
+    // MARK: - Public Methods
+    
     // TODO: - Remove this static method and require scheme in init
+    /// Set the URL Scheme that you have registered in your Info.plist.
     public static func setReturnURLScheme(_ returnURLScheme: String?) {
         self.scheme = returnURLScheme
     }
     
+    /// Handle completion of the popup flow by calling this method from either
+    /// your scene:openURLContexts: scene delegate method or
+    /// your application:openURL:sourceApplication:annotation: app delegate method.
     @objc(openURL:)
     public static func open(url: URL) -> Bool {
         if let returnBlock {
@@ -75,12 +68,35 @@ public class POPPopupBridgeSwift: NSObject, WKScriptMessageHandler, SFSafariView
         }
     }
     
-    // URL from SceneDelegate --> JS, to be injected back into merchant WebView
-    func parseResponseFromPopUpToPassBacktoWebView(url: URL) -> String? {
-        guard let urlComponents = URLComponents.init(url: url, resolvingAgainstBaseURL: false),
+    // MARK: - Internal Methods
+    
+    /// Injects custom JavaScript into the merchant's webpage.
+    /// - Parameter scheme: the url scheme provided by the merchant
+    private func configureWebView(scheme: String) {
+        webView.configuration.userContentController.add(self, name: messageHandlerName)
+        
+        let javascript = PopupBridgeUserScript(
+            scheme: scheme,
+            scriptMessageHandlerName: messageHandlerName,
+            host: hostName
+        ).rawJavascript
+        
+        let script = WKUserScript(
+            source: javascript,
+            injectionTime: .atDocumentStart,
+            forMainFrameOnly: true
+        )
+        webView.configuration.userContentController.addUserScript(script)
+    }
+    
+    /// Constructs custom JavaScript to be injected into the merchant's WKWebView, based on redirectURL details from the SFSafariViewController pop-up result.
+    /// - Parameter url: returnURL from the result of the SFSafariViewController being dismissed. Provided via the merchant's AppDelegate or SceneDelegate.
+    /// - Returns: JavaScript formatted completion.
+    private func constructJavaScriptCompletionResult(returnURL: URL) -> String? {
+        guard let urlComponents = URLComponents.init(url: returnURL, resolvingAgainstBaseURL: false),
               let scheme = POPPopupBridgeSwift.scheme,
               urlComponents.scheme?.caseInsensitiveCompare(scheme) == .orderedSame,
-              urlComponents.host?.caseInsensitiveCompare(kPOPURLHost) == .orderedSame
+              urlComponents.host?.caseInsensitiveCompare(hostName) == .orderedSame
         else {
             return nil
         }
@@ -91,7 +107,7 @@ public class POPPopupBridgeSwift: NSObject, WKScriptMessageHandler, SFSafariView
         
         var payloadDictionary: [String: Any] = [:]
         payloadDictionary["path"] = urlComponents.path
-        payloadDictionary["queryItems"] = url.queryDictionary
+        payloadDictionary["queryItems"] = returnURL.queryDictionary
         payloadDictionary["hash"] = urlComponents.fragment
         
         do {
@@ -108,17 +124,25 @@ public class POPPopupBridgeSwift: NSObject, WKScriptMessageHandler, SFSafariView
         
     }
     
-    func dismissSafariViewController() {
+    private func dismissSafariViewController() {
         if let safariViewController {
-            if delegate.responds(to: #selector(POPPopupBridgeDelegate.popupBridge(_:requestsDismissalOf:))) {
-                delegate.popupBridge(self, requestsDismissalOfViewController: safariViewController)
+            delegate.popupBridge(self, requestsDismissalOfViewController: safariViewController)
+        }
+    }
+    
+    private func injectWebView(webView: WKWebView, withJavaScript script: String) {
+        webView.evaluateJavaScript(script) { _, error in
+            if let error {
+                NSLog("Error: PopupBridge requires onComplete callback. Details: %@", error.localizedDescription)
             }
         }
     }
     
-    // MARK: - SFSafariViewControllerDelegate
+    // TODO: - Make below methods private by moving protocol conformances into helper classes.
     
-    // User clicked "Done"
+    // MARK: - SFSafariViewControllerDelegate conformance
+    
+    /// Called when the user exits the pop-up (SFSafariViewController) by clicking "Done"
     public func safariViewControllerDidFinish(_ controller: SFSafariViewController) {
         let script = """
             if (typeof window.popupBridge.onCancel === 'function') {\
@@ -131,7 +155,6 @@ public class POPPopupBridgeSwift: NSObject, WKScriptMessageHandler, SFSafariView
     }
     
     // MARK: - WKScriptMessageHandler conformance
-    // Receives messages from JS
     
     public func userContentController(_ userContentController: WKUserContentController, didReceive message: WKScriptMessage) {
         if message.name == kPOPScriptMessageHandlerName {
@@ -139,11 +162,8 @@ public class POPPopupBridgeSwift: NSObject, WKScriptMessageHandler, SFSafariView
             if let urlString = params["url"] as! String?,
                let url = URL(string: urlString) {
                 dismissSafariViewController()
-                                
-//                 do we need to do this responds to selector thing?
-//                if delegate.responds(to: #selector(POPPopupBridgeDelegate.popupBridge(_:willOpen:))) {
+                
                 delegate.popupBridge?(self, willOpenURL: url)
-//                }
                 
                 safariViewController = SFSafariViewController(url: url)
                 safariViewController?.delegate = self
@@ -152,20 +172,9 @@ public class POPPopupBridgeSwift: NSObject, WKScriptMessageHandler, SFSafariView
             }
             
             // TODO - this parsing is awful
-            if let name = (params as? [String: [String: String]])?["message"]?["name"],
-               delegate.responds(to: #selector(POPPopupBridgeDelegate.popupBridge(_:receivedMessage:data:))) {
+            if let name = (params as? [String: [String: String]])?["message"]?["name"] {
                 let data = (params as? [String: [String: String]])?["message"]?["data"]
                 delegate.popupBridge?(self, receivedMessage: name, data: data)
-            }
-        }
-    }
-    
-    // MARK: - Helpers
-    
-    func injectWebView(webView: WKWebView, withJavaScript script: String) {
-        webView.evaluateJavaScript(script) { _, error in
-            if let error {
-                NSLog("Error: PopupBridge requires onComplete callback. Details: %@", error.localizedDescription)
             }
         }
     }