+++ Sources/PopupBridge/PopupBridgeUserScript.swift
@@ -1,6 +1,6 @@
 import Foundation
 
-struct UserScript {
+struct PopupBridgeUserScript {
     let scheme: String
     let scriptMessageHandlerName: String
     let host: String